services/event-emitter-mqtt/requirements.txt
paho-mqtt
mongoengine
uuid
datetime